feat: Add logic to ignore comments in toBeEmptyDOMElement
After a small discussion move logic from custom matcher to the existing toBeEmptyDOMElement.
